# Add `index_add` benchmark