--- a/test/utils/test_scatter.py
+++ b/test/utils/test_scatter.py
@@ -1,3 +1,5 @@
+from itertools import product
+
 import pytest
 import torch
 
@@ -84,7 +86,7 @@ def test_scatter_argmax(device):
 if __name__ == '__main__':
     # Insights on GPU:
     # ================
-    # * "sum": Prefer `index_add_` implementation
+    # * "sum": Prefer `scatter_add_` implementation
     # * "mean": Prefer manual implementation via `scatter_add_` + `count`
     # * "min"/"max":
     #   * Prefer `scatter_reduce_` implementation without gradients
@@ -93,7 +95,7 @@ if __name__ == '__main__':
     #
     # Insights on CPU:
     # ================
-    # * "sum": Prefer `index_add_` implementation
+    # * "sum": Prefer `scatter_add_` implementation
     # * "mean": Prefer manual implementation via `scatter_add_` + `count`
     # * "min"/"max": Prefer `scatter_reduce_` implementation
     # * "mul" (probably not worth branching for this):
@@ -101,60 +103,72 @@ if __name__ == '__main__':
     #   * Prefer `torch_sparse` implementation with gradients
     import argparse
 
+    from torch_geometric.typing import WITH_TORCH_SCATTER, torch_scatter
+
     parser = argparse.ArgumentParser()
     parser.add_argument('--device', type=str, default='cuda')
     parser.add_argument('--backward', action='store_true')
-    parser.add_argument(
-        '--aggr', type=str, default='all', help=
-        "Specify a specific aggr to benchmark or multiple seperated by commas")
+    parser.add_argument('--aggr', type=str, default='all')
     args = parser.parse_args()
 
-    for num_nodes in [1000, 2000, 4000, 8000, 16000, 32000, 64000, 128000]:
+    num_nodes_list = [4_000, 8_000, 16_000, 32_000, 64_000]
+
+    if args.aggr == 'all':
+        aggrs = ['sum', 'mean', 'min', 'max', 'mul']
+    else:
+        aggrs = args.aggr.split(',')
+
+    def pytorch_scatter(x, index, dim_size, reduce):
+        if reduce == 'min' or reduce == 'max':
+            reduce = f'a{aggr}'  # `amin` or `amax`
+        elif reduce == 'mul':
+            reduce = 'prod'
+        out = x.new_zeros(dim_size, x.size(-1))
+        include_self = reduce in ['sum', 'mean']
+        index = index.view(-1, 1).expand(-1, x.size(-1))
+        out.scatter_reduce_(0, index, x, reduce, include_self=include_self)
+        return out
+
+    def pytorch_index_add(x, index, dim_size, reduce):
+        if reduce != 'sum':
+            raise NotImplementedError
+        out = x.new_zeros(dim_size, x.size(-1))
+        out.index_add_(0, index, x)
+        return out
+
+    def own_scatter(x, index, dim_size, reduce):
+        return torch_scatter.scatter(x, index, dim=0, dim_size=num_nodes,
+                                     reduce=reduce)
+
+    def optimized_scatter(x, index, dim_size, reduce):
+        return scatter(x, index, dim=0, dim_size=dim_size, reduce=reduce)
+
+    for aggr, num_nodes in product(aggrs, num_nodes_list):
         num_edges = num_nodes * 50
-        print("Benchmarking w/ (num_nodes, num_edges) =",
-              (num_nodes, num_edges))
+        print(f'aggr: {aggr}, #nodes: {num_nodes}, #edges: {num_edges}')
+
         x = torch.randn(num_edges, 64, device=args.device)
         index = torch.randint(num_nodes, (num_edges, ), device=args.device)
 
-        from torch_geometric.typing import WITH_TORCH_SCATTER
+        funcs = [pytorch_scatter]
+        func_names = ['PyTorch scatter_reduce']
+
+        if aggr == 'sum':
+            funcs.append(pytorch_index_add)
+            func_names.append('PyTorch index_add')
+
         if WITH_TORCH_SCATTER:
-            import torch_scatter
+            funcs.append(own_scatter)
+            func_names.append('torch_scatter')
 
-        def pytorch_scatter(x, index, dim_size, reduce):
-            if reduce == 'min' or reduce == 'max':
-                reduce = f'a{aggr}'  # `amin` or `amax`
-            elif reduce == 'mul':
-                reduce = 'prod'
-            out = x.new_zeros((dim_size, x.size(-1)))
-            include_self = reduce in ['sum', 'mean']
-            index = index.view(-1, 1).expand(-1, x.size(-1))
-            out.scatter_reduce_(0, index, x, reduce, include_self=include_self)
-            return out
+        funcs.append(optimized_scatter)
+        func_names.append('Optimized PyG Scatter')
 
-        def own_scatter(x, index, dim_size, reduce):
-            return torch_scatter.scatter(x, index, dim=0, dim_size=num_nodes,
-                                         reduce=reduce)
-
-        def optimized_scatter(x, index, dim_size, reduce):
-            return scatter(x, index, dim=0, dim_size=dim_size, reduce=reduce)
-
-        if args.aggr == 'all':
-            aggrs = ['sum', 'mean', 'min', 'max', 'mul']
-        else:
-            aggrs = args.aggr.split(',')
-        for aggr in aggrs:
-            print(f'Aggregator: {aggr}')
-            funcs = [pytorch_scatter, optimized_scatter]
-            func_names = ['PyTorch Scatter', 'Optimized PyG Scatter']
-            if WITH_TORCH_SCATTER:
-                funcs.append(own_scatter)
-                func_names.append('torch_scatter')
-
-            benchmark(
-                funcs=funcs,
-                func_names=func_names,
-                args=(x, index, num_nodes, aggr),
-                num_steps=100 if args.device == 'cpu' else 1000,
-                num_warmups=50 if args.device == 'cpu' else 500,
-                backward=args.backward,
-            )
+        benchmark(
+            funcs=funcs,
+            func_names=func_names,
+            args=(x, index, num_nodes, aggr),
+            num_steps=100 if args.device == 'cpu' else 1000,
+            num_warmups=50 if args.device == 'cpu' else 500,
+            backward=args.backward,
+        )