using: https://github.com/puririshi98/rgcn_pyg_lib_forward_bench/blob/main/scatter_v_index_bench.py
I find that index_add is faster
```
original implementation takes 3.62396240234375e-07 s/iter
new implementation takes 2.765655517578125e-07 s/iter
```

however using test/utils/test_scatter.py
I am seeing the opposite result
```
Benchmarking w/ (num_nodes, num_edges) = (1000, 50000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 0.0512s   |
| Optimized PyG Scatter | 0.0699s   |
+-----------------------+-----------+
Benchmarking w/ (num_nodes, num_edges) = (2000, 100000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 0.0752s   |
| Optimized PyG Scatter | 0.1124s   |
+-----------------------+-----------+
Benchmarking w/ (num_nodes, num_edges) = (4000, 200000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 0.1227s   |
| Optimized PyG Scatter | 0.2087s   |
+-----------------------+-----------+
Benchmarking w/ (num_nodes, num_edges) = (8000, 400000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 0.2408s   |
| Optimized PyG Scatter | 0.4129s   |
+-----------------------+-----------+
Benchmarking w/ (num_nodes, num_edges) = (16000, 800000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 0.5714s   |
| Optimized PyG Scatter | 0.8529s   |
+-----------------------+-----------+
Benchmarking w/ (num_nodes, num_edges) = (32000, 1600000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 1.3682s   |
| Optimized PyG Scatter | 1.8042s   |
+-----------------------+-----------+
Benchmarking w/ (num_nodes, num_edges) = (64000, 3200000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 3.2057s   |
| Optimized PyG Scatter | 3.7839s   |
+-----------------------+-----------+
Benchmarking w/ (num_nodes, num_edges) = (128000, 6400000)
Aggregator: sum
+-----------------------+-----------+
| Name                  | Forward   |
|-----------------------+-----------|
| PyTorch Scatter       | 6.9272s   |
| Optimized PyG Scatter | 7.8657s   |
+-----------------------+-----------+
```