@rusty1s 2023-06-24
I think you are missing a `cuda.synchronize()` in your own benchmark.